+++ ubun2_be/src/main/java/kr/or/kosa/ubun2_be/domain/member/controller/MemberController.java
@@ -1,10 +1,7 @@
 package kr.or.kosa.ubun2_be.domain.member.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
-import kr.or.kosa.ubun2_be.domain.member.dto.EmailAuthenticationRequest;
-import kr.or.kosa.ubun2_be.domain.member.dto.EmailRequest;
 import kr.or.kosa.ubun2_be.domain.member.dto.MemberSignUpRequest;
-import kr.or.kosa.ubun2_be.domain.member.service.EmailService;
 import kr.or.kosa.ubun2_be.domain.member.service.MemberService;
 import kr.or.kosa.ubun2_be.global.dto.ResponseDto;
 import lombok.RequiredArgsConstructor;
@@ -15,25 +12,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/members")
 public class MemberController {
     private final MemberService memberService;
-    private final EmailService emailService;
 
     @Operation(summary = "회원 회원가입")
     @PostMapping("/signup")
     public ResponseDto<?> createMember(@RequestBody MemberSignUpRequest memberSignUpRequest) {
         memberService.createMember(memberSignUpRequest);
         return ResponseDto.ok(null, "정상출력 데이터");
-    }
-    @Operation(summary = "인증번호 발송")
-    @PostMapping("/auth/send")
-    public ResponseDto<?> sendEmail(@RequestBody EmailRequest emailRequest) {
-        emailService.sendEmail(emailRequest);
-        return ResponseDto.ok(null, "정상출력 데이터");
-    }
-    @Operation(summary = "인증번호 검증")
-    @PostMapping("/auth")
-    public ResponseDto<?> checkAuthenticationNumber(@RequestBody EmailAuthenticationRequest emailAuthenticationRequest) {
-        boolean isValid = emailService.validateAuthenticationNumber(emailAuthenticationRequest);
-        return ResponseDto.ok(isValid, "정상출력 데이터");
     }
 
 }
