+++ ubun2_be/src/main/java/kr/or/kosa/ubun2_be/global/auth/service/EmailService.java
@@ -1,0 +1,9 @@
+package kr.or.kosa.ubun2_be.global.auth.service;
+
+import kr.or.kosa.ubun2_be.global.auth.dto.EmailAuthenticationRequest;
+import kr.or.kosa.ubun2_be.global.auth.dto.EmailRequest;
+
+public interface EmailService {
+    void sendEmail(EmailRequest emailRequest);
+    boolean validateAuthenticationNumber(EmailAuthenticationRequest emailAuthenticationRequest);
+}
