+++ ubun2_be/src/main/java/kr/or/kosa/ubun2_be/global/config/CorsMvcConfig.java
@@ -1,4 +1,4 @@
-package kr.or.kosa.ubun2_be.global.auth.config;
+package kr.or.kosa.ubun2_be.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
